Move hand washing method files to right paths
